todo cleanup - prevent audit channel deletion

+++ src/main/java/com/flightstats/hub/dao/encryption/AuditChannelService.java
@@ -155,7 +155,9 @@ public class AuditChannelService implements ChannelService {
 
     @Override
     public boolean delete(String channelName) {
-        //todo - gfm - 5/15/14 - prevent deletion of _audit ?
+        if (isAuditChannel(channelName)) {
+            throw new ForbiddenRequestException("Audit Channels can not be deleted.");
+        }
         return channelService.delete(channelName);
     }
